add config endpoints
