+++ examples/upgrade-contract/index.js
@@ -9,6 +9,10 @@ async function main() {
     const creds = { apiKey: process.env.ADMIN_API_KEY, apiSecret: process.env.ADMIN_API_SECRET };
     const client = new PlatformClient(creds);
 
+    // Get approval process for upgrade on Goerli
+    const config = await client.Upgrade.getApprovalProcess("goerli");
+    console.log(config);
+
     const upgrade = await client.Upgrade.upgrade({
         proxyAddress: "0xABC1234...",
         proxyAdminAddress: "0xDEF1234...",
