+++ packages/deploy/src/models/deployment-config.ts
@@ -1,0 +1,14 @@
+import { Network } from 'defender-base-client';
+
+export interface DeploymentConfigCreateRequest {
+  relayerId: string;
+  stackResourceId?: string;
+}
+
+export interface DeploymentConfigResponse {
+  deploymentConfigId: string;
+  relayerId: string;
+  network: Network;
+  createdAt: string;
+  stackResourceId?: string;
+}
